Implement various Python scripts for calculations, user input handling, and data processing. - Added a payroll calculator for workers based on categories and hours worked. - Enhanced time conversion script to calculate total seconds from days, hours, minutes, and seconds. - Developed a reverse time conversion script to display days, hours, minutes, and seconds from a total seconds input. - Created a rectangle area and perimeter calculator. - Implemented distance conversion between kilometers and other units. - Developed a sphere volume and surface area calculator. - Created a program to generate figures based on user-defined levels. - Added a patent verification system with user input handling. - Implemented a PDF generator class for creating formatted documents. - Updated Docker configuration for database service. - Added ATM simulation scripts for cash withdrawal with bill denomination calculations. - Enhanced user interaction scripts for client spending tracking.

+++ 20230328/ejer4.py
@@ -6,7 +6,7 @@ KM_MILLAS = 0.621371
 KM_PULGADAS = 39370.1
 KM_METROS = 1000
 
-kilometros = int(input('Ingresar la cantidad en kilometros: '))
+kilometros = float(input('Ingresar la cantidad en kilometros: '))
 
 yardas = kilometros * KM_YARDAS
 millas = kilometros * KM_MILLAS
